[#84] Added progress in work on source detection
- updated util for source-type's detection
- done small refactoring
