+++ src/modules/podcast/models.py
@@ -1,4 +1,3 @@
-import enum
 import uuid
 from hashlib import md5
 from datetime import datetime
@@ -12,25 +11,7 @@ from core import settings
 from core.database import ModelBase
 from common.models import ModelMixin
 from common.db_utils import EnumTypeColumn
-from modules.providers.utils import SourceType
-
-
-class EpisodeStatus(enum.Enum):
-    NEW = "new"
-    DOWNLOADING = "downloading"
-    PUBLISHED = "published"
-    ARCHIVED = "archived"
-    ERROR = "error"
-
-    DL_PENDING = "pending"
-    DL_EPISODE_DOWNLOADING = "episode_downloading"
-    DL_EPISODE_POSTPROCESSING = "episode_postprocessing"
-    DL_EPISODE_UPLOADING = "episode_uploading"
-    DL_COVER_DOWNLOADING = "cover_downloading"
-    DL_COVER_UPLOADING = "cover_uploading"
-
-    def __str__(self):
-        return self.value
+from common.enums import SourceType, EpisodeStatus
 
 
 class Podcast(ModelBase, ModelMixin):
